+++ ckanext/opendata_theme/base/helpers.py
@@ -2,6 +2,7 @@ import ast
 import logging
 
 from ckan.plugins import toolkit
+from ckan.plugins.toolkit import config
 
 from ckanext.opendata_theme.base.compatibility_controller import BaseCompatibilityController
 from ckanext.opendata_theme.opengov_custom_homepage.constants import CUSTOM_NAMING
@@ -95,6 +96,14 @@ def package_tracking_summary(package):
         logger.debug("[opendata_theme] Error getting dataset tracking_summary")
         return {}
     return tracking_summary
+
+
+def get_group_alias():
+    return str(config.get('ckan.group_alias', 'Group'))
+
+
+def get_organization_alias():
+    return str(config.get('ckan.organization_alias', 'Organization'))
 
 
 def get_custom_name(key, default_name):
