+++ ckanext/opendata_theme/opengov_custom_footer/plugin/__init__.py
@@ -12,7 +12,7 @@ else:
     from ckan.lib.helpers import literal
 from ckanext.opendata_theme.opengov_custom_footer.common_controller import CustomFooterCommonController
 from ckanext.opendata_theme.opengov_custom_footer.constants import CONFIG_SECTION
-from ckanext.opendata_theme.base.template_helpers import version_builder
+from ckanext.opendata_theme.base.utils import version_builder
 
 
 class Opendata_ThemePlugin(MixinPlugin):
