+++ ckanext/opendata_theme/opengov_custom_header/templates/default_header.html
@@ -15,26 +15,35 @@
   {% set group_index_route = 'group_index' %}
 {% endif %}
 
-{% block header_wrapper %}
+{% block header_wrapper %} {% block header_account %}
+<div class="account-masthead">
+  <div class="container">
+    {% block header_account_container_content %}
+      {{ super() }}
+    {% endblock %}
+  </div>
+</div>
+{% endblock %}
 <header class="navbar navbar-static-top masthead">
   {% block header_debug %}
     {{ super() }}
   {% endblock %}
   <div class="container">
-    <button data-target=".nav-collapse" data-toggle="collapse" class="btn btn-navbar" type="button">
-      <span class="fa fa-bars"></span>
-    </button>
-    <hgroup class="{{ g.header_class }} pull-left">
+    <div class="navbar-right">
+      <button data-target="#main-navigation-toggle" data-toggle="collapse" class="navbar-toggle collapsed" type="button" aria-label="expand or collapse" aria-expanded="false">
+        <span class="sr-only">{{ _('Toggle navigation') }}</span>
+        <span class="fa fa-bars"></span>
+      </button>
+    </div>
+    <hgroup class="{{ g.header_class }} navbar-left">
       {% block header_logo %}
         {{ super() }}
       {% endblock %}
     </hgroup>
-
-    <div class="nav-collapse collapse container">
-
+    <div class="collapse navbar-collapse" id="main-navigation-toggle">
       {% block header_site_navigation %}
-        <nav class="section navigation">
-          <ul class="nav nav-pills">
+      <nav class="section navigation">
+        <ul class="nav nav-pills">
             {% block header_site_navigation_tabs %}
               {{ h.build_nav_main(
                 (search_route, _('Datasets')),
@@ -46,33 +55,12 @@
             <li>
                 <a href="{{ h.url_for('/about') }}" title="{{ _('About') }}">{{ _('About') }}</a>
             </li>
-          </ul>
-        </nav>
-      {% endblock %}
-      {% block header_account %}
-      <div class="account-masthead">
-          {% block header_account_container_content %}
-            {% if c.userobj %}
-              <div class="account avatar authed" data-module="me" data-me="{{ c.userobj.id }}">
-                <ul class="unstyled">
-                  {% block header_account_logged %}
-                    {{ super() }}
-                  {% endblock %}
-                </ul>
-              </div>
-            {% else %}
-              <nav class="account not-authed">
-                <ul class="unstyled">
-                  {% block header_account_notlogged %}
-                    {{ super() }}
-                  {% endblock %}
-                </ul>
-              </nav>
-            {% endif %}
           {% endblock %}
-      </div>
+        </ul>
+      </nav>
+      {% block header_site_search %}
+        {{ super() }}
       {% endblock %}
-
     </div>
   </div>
 </header>
