+++ ckanext/opendata_theme/opengov_custom_header/templates/page.html
@@ -4,6 +4,6 @@
     {% if layout_type=='compressed' %}
         {% include "compressed_header.html" %}
     {% else %}
-        {% include "header.html" %}
+        {% include "default_header.html" %}
     {% endif %}
 {% endblock %}
