+++ requirements.txt
@@ -1,3 +1,3 @@
 wcag_contrast_ratio==0.9
-webcolors==1.10
+webcolors==1.12
 bleach
